Previous commit was slightly off in problem number, this is problem b, previous was a.
This problem encompasses encoding quantum gates dependent on learnable
parameters. The number of such should be arbitrary (should proobably be
even numbers however, due to the pairwise coupling of the gates.

I suspect there is a mismatch between my understainding and the expected
one, as the number of trainable parameters ϴ should be arbitraty, while
the number of features are referred to as the "p first" parameters. The
number of parameters ϴ and p differs in the examples given as well.

So far, I think I can only acknowledge this and try to implement the
code as best I can and anticipate changes.