I struggled with starting on the already qml, so I wrote an entirely separate one. don't know why, but it's there now. so.
the "main.py" file has a qml class that takes in external functions as
ansatz and encoder, running them in a model class for each epoch to
construct the circuit and run it. This allows for flexibility irt.
inputting the different ansatzes and encoders later on. One issue with
this is that I think the inputs to the functions can become somewhat
excessive since they don't have access to the "self" objects, which is
sort of why we made the class to begin with.

Another alternative, which I think might work better might be to
implement the encoders and ansatzes as separate entities in the code,
with a dictionary for each and then setting the desired set for each. so
something along the lines of

class Bla:
    def __init__(self, ..., encoder, ansatz):
	self...
	.
	.
	.
	self.encoder = encoders[encoder]
	self.ansats = ansatzes[ansatz]

with the "encoder" and "ansatz" being a string for the name of the
specific version implemented and storing the dictionaries of encoders
and ansatzes implemented, checking upon initialization/setting the
encoder/ansatz. I think then that the encoders and ansatzes should be
based on NOT importing any data, and rather using the "self" domain to
access the necessary objects, possibly with the exception of the
circuit, which they could take in and modify.