Implemented importing and structuring the iris flower data set into the code. Currently, the class handles 1 sample and so expansions would allow for running over several. There is a question of whether or not the code should run the current θ variables on every sample and adjust the average, or if they should be adjusted properly to the first, seconde, third... and then possibly find the average of each set of θ's later.
The basic ansatz and encoder now works with radians rather than just
random numbers (scaling the random initial theta's with 2π and encoding
the features with 2π).

The loss function is still MSE which needs to be rectified to the cross
entropy, but I don't have the energy to figure that one out right now.