Loss function
Attempt at adding in the proper loss function for logistic regression.
Currently it is the cross entropy. It is not scaled to any number, and
will likely not work for enything besides the case of using a single
sample. I did not include the proper derivative of theta in the
function, but after thinking a bit, I think this is the theta gradient
we find manually. Everything needs work, but it runs and it's a
different loss than plain MSE.