cleanup of maths and ansatz
In meeting with TA's. Went through some maths irt. the calculation of
the mean loss and correction of the model weights and biases. changed
the encoding and ansatz for the input and tweaked the code some. overall
performance seems in line with what we want for now.

next steps are likely experimenting with encoding and ansatzes.